duyet

+++ src/java/webservices/UserAPI.java
@@ -6,16 +6,13 @@
 package webservices;
 
 import com.google.gson.Gson;
+import common.DungChung;
 import common.DungChung.ReturnMessage;
 import dao.IRole;
 import dao.UserDAO;
 import entities.Users;
-import java.util.Enumeration;
 import java.util.List;
-import javax.annotation.security.PermitAll;
-import javax.annotation.security.RolesAllowed;
 import javax.ejb.Stateless;
-import javax.servlet.http.HttpServletRequest;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -26,7 +23,6 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
 import model.CurrentUser;
 
 /**
@@ -46,50 +42,76 @@ public class UserAPI extends BaseAPI {
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     public String getAll(@Context HttpHeaders httpHeader) {
-        List<String> auths = httpHeader.getRequestHeader("authorization");
-        if (auths != null) {
-            CurrentUser cu = getCurrentUser(auths.get(0));
-            if (cu != null) {
+        CurrentUser cu = getCurrentUser(httpHeader);
+        if (cu != null) {
+            List<String> roles = cu.getRoles();
+            if (IRole.isRole(roles, IRole.LEVEL.LOW)) {
                 Gson g = new Gson();
-                db.setCurrentUser(cu);
                 String data = g.toJson(db.getData());
                 return data;
             }
-            return "Not login!";
+            return DungChung.MESSAGE.NOT_AUTHORIZATION;
         }
-        return "Not authorization!";
+        return DungChung.MESSAGE.NOT_LOGIN;
     }
 
     @Path("/{id}")
     @GET
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.TEXT_PLAIN)
-    public String getById(@PathParam("id") String id) {
-        Gson g = new Gson();
-        String data = g.toJson(db.getById(id));
-        return data;
+    public String getById(@Context HttpHeaders httpHeader, @PathParam("id") String id) {
+        CurrentUser cu = getCurrentUser(httpHeader);
+        if (cu != null) {
+            List<String> roles = cu.getRoles();
+            if (IRole.isRole(roles, IRole.LEVEL.LOW)) {
+                Gson g = new Gson();
+                String data = g.toJson(db.getById(id));
+                return data;
+            }
+            return DungChung.MESSAGE.NOT_AUTHORIZATION;
+        }
+        return DungChung.MESSAGE.NOT_LOGIN;
+
     }
 
     @POST
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public String insert(String entity) {
-        Gson g = new Gson();
-        Users p = g.fromJson(entity, Users.class);
-        ReturnMessage msg = db.setData(p);
-        String data = g.toJson(msg);
-        return data;
+    public String insert(@Context HttpHeaders httpHeader, String entity) {
+        CurrentUser cu = getCurrentUser(httpHeader);
+        if (cu != null) {
+            List<String> roles = cu.getRoles();
+            if (IRole.isRole(roles, IRole.LEVEL.MEDIUM)) {
+                db.setCurrentUser(cu);
+                Gson g = new Gson();
+                Users p = g.fromJson(entity, Users.class);
+                ReturnMessage msg = db.setData(p);
+                String data = g.toJson(msg);
+                return data;
+            }
+            return DungChung.MESSAGE.NOT_AUTHORIZATION;
+        }
+        return DungChung.MESSAGE.NOT_LOGIN;
     }
 
     @DELETE
     @Path("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.TEXT_PLAIN)
-    public String delete(@PathParam("id") String id) {
-        Gson g = new Gson();
-        ReturnMessage msg = db.removeData(id);
-        String data = g.toJson(msg);
-        return data;
+    public String delete(@Context HttpHeaders httpHeader, @PathParam("id") String id) {
+        CurrentUser cu = getCurrentUser(httpHeader);
+        if (cu != null) {
+            List<String> roles = cu.getRoles();
+            if (IRole.isRole(roles, IRole.LEVEL.MEDIUM)) {
+                Gson g = new Gson();
+                ReturnMessage msg = db.removeData(id);
+                String data = g.toJson(msg);
+                return data;
+            }
+            return DungChung.MESSAGE.NOT_AUTHORIZATION;
+        }
+        return DungChung.MESSAGE.NOT_LOGIN;
+
     }
 
 }
